Fix the bytebeat bot to work from stdin so we can queue easily
The current configuration of our bot uses `head` and `sed` to manage a queue of
posts, and then feeds the items of the queue to the bot on stdin.

This also makes it compatible with FFmpeg 2.8.11, since that's what we have on
our server. FFmpeg 2.8 doesn't detect PPMs from the contents, so we need to call
our video input video.ppm.

Fixes #37