Improve random generation
The random generator now produces bytebeats in the form "t-like osci mul"

Where a "t like" formula is one that has a period of approximately t,
an "osci" formula has a low period (usually between t>>4 and t>>13).

These formulas are made up of multiple bytebeat units, which are valid bytebeat
formulas that end with exactly one value on the stack. These are useful as
they can be substituted in place where a variable or constant is. For example,
one may subsistue the unit `t` in `t*(t>>4 & t>>8)` with the unit `t>>t`,
producing `t>>t*(t>>4 & t>>8)`